Propose a better views implementation
